#172 add postGeneralInfo in GenreralInfoService
